feat: add rewind command

+++ src/event_scanner.rs
@@ -59,7 +59,7 @@ impl EventScanner {
 
     /// Configures how many blocks are read per epoch during a historical sync.
     #[must_use]
-    pub fn with_blocks_read_per_epoch(mut self, blocks_read_per_epoch: usize) -> Self {
+    pub fn with_blocks_read_per_epoch(mut self, blocks_read_per_epoch: u32) -> Self {
         self.block_range_scanner =
             self.block_range_scanner.with_blocks_read_per_epoch(blocks_read_per_epoch);
         self
